Updated VariableSymbol namespace and moved it to its own folder
