+++ Language/CodeAnalysis/Binding/BoundStatement.cs
@@ -1,4 +1,6 @@
-﻿namespace Language.CodeAnalysis.Binding
+﻿using Language.CodeAnalysis.Symbols;
+
+namespace Language.CodeAnalysis.Binding
 {
     internal abstract class BoundStatement : BoundNode
     {
